no ~> not
I happened to have a similar thought. I figured anything regarding the President's daughter and anything out-of-the-ordinary would indeed be newsworthy. Curious that we never again heard about that night from the press. Not even an allusion to the President already having to deal with bad press or the like...

Ah well, it is a liberal fantasy after all.